Working without config

+++ client.go
@@ -53,7 +53,7 @@ func (c *client) readPump() {
 		c.ws.Close()
 	}()
 
-	c.ws.SetReadLimit(maxMessageSize)
+	c.ws.SetReadLimit(int64(maxMessageSize))
 	c.ws.SetReadDeadline(time.Now().Add(pongWait))
 	c.ws.SetPongHandler(func(string) error {
 		c.ws.SetReadDeadline(time.Now().Add(pongWait))
